requirements.txt
boto3
tqdm
environs
bson
shortuuid
python-json-logger
pydantic
requests
omegaconf
speechbrain
transformers
torchvision
denoiser
webrtcvad
sphfile
git+https://github.com/openai/whisper.git
git+https://github.com/facebookresearch/demucs#egg=demucs
git+https://github.com/pyannote/pyannote-audio.git